requirements.txt
asgiref==3.2.5
certifi==2019.11.28
chardet==3.0.4
django-cloudinary-storage==0.2.3
cloudinary==1.20.0    
Pillow==6.2.2
Django==2.2.10
gunicorn==20.0.4
whitenoise==5.0.1
django-crispy-forms==1.7.2
pytz==2018.9
sqlparse==0.3.0
python-decouple==3.3
psycopg2-binary==2.8.3
requests==2.23.0
dj-database-url==0.5.0    
django-allauth==0.41.0
easy-thumbnails==2.7  
hashids==0.8.4
django-ckeditor==5.9.0
django-allauth==0.41.0
